Added a new language - German
Added German and some unit tests to verify everything is working correctly. Fixed a couple of other bugs such as the background view of the acknowledgements view controller to systemBackground to allow for light/dark mode. Also changed the hardMode icons to a transparent icon with a background tint color instead of creating a single icon for each language